+++ src/test/java/fr/marcwrobel/jbanking/internal/SerializationUtils.java
@@ -25,12 +25,17 @@ public class SerializationUtils {
 
   /**
    * Check whether the given class is serializable.
-   * 
+   *
    * @param clazz a non-null {@link Class}
    * @return {@code true} if the class is serializable, {@code false} otherwise
    */
   public static boolean isSerializable(Class<?> clazz) {
-    boolean serializable = clazz.isPrimitive() || clazz.isInterface() || Serializable.class.isAssignableFrom(clazz);
+    if (clazz.isEnum()) {
+      return true; // see https://docs.oracle.com/javase/specs/jls/se8/html/jls-8.html#jls-8.9
+    }
+
+    boolean serializable = clazz.isPrimitive() || clazz.isInterface()
+        || Serializable.class.isAssignableFrom(clazz);
     if (!serializable) {
       return false;
     }